% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{assign_to_cluster}
\alias{assign_to_cluster}
\title{Assign new data to existing clusters}
\usage{
assign_to_cluster(data, data_centers, data_scales, cluster_centers)
}
\arguments{
\item{data}{dataframe output of \code{\link{process_landis}}}

\item{data_centers}{center parameters for scaled data}

\item{data_scales}{scale parameters for scaled data}

\item{cluster_centers}{matrix of center values for clusters}
}
\value{
vector of integer cluster assignments
}
\description{
Assign new data to existing clusters
}
\examples{
assign_to_cluster(
  data = process_landis(data.table::fread(test_landis_prediction_file)),
  data_centers = test_clusters$data_centers,
  data_scales = test_clusters$data_scales,
  cluster_centers = test_clusters$kmeans$center
)
}
