% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{backcast_landis_to_treelists}
\alias{backcast_landis_to_treelists}
\title{Backcast Landis to treelists}
\usage{
backcast_landis_to_treelists(new_data, lookup, n_clusters = 50, treelist_dir)
}
\arguments{
\item{new_data}{dataframe object that will get backcasted treelists, output
of \code{\link{process_landis}}}

\item{lookup}{dataframe output of \code{\link{process_landis}} from which
matches will be drawn.}

\item{n_clusters}{integer number of clusters for grouping the pool of input
pixels in \code{lookup}.}

\item{treelist_dir}{file path to directory containing treelist csv files
corresponding to the pool of pixels in the lookup dataset.}
}
\value{
list containing backcasted treelists for pixels in \code{new_data},
a dataframe with matching accuracy statistics (\code{comp_stats}), and a
dataframe with original and matched values of biomass by species and age
class for \code{new_data} (\code{comp_frame})
}
\description{
Backcast Landis to treelists
}
\examples{
backcasted <- backcast_landis_to_treelists(
  new_data = process_landis(data.table::fread(test_landis_prediction_file)),
  lookup = test_landis,
  n_clusters = 20,
  treelist_dir = test_landis_treelist_dir
)
}
