% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canopy_cover_table}
\alias{canopy_cover_table}
\title{A dataframe with estimates of canopy cover proportion with bapa and tpa
observations from FIA data in the Sierra Nevada region}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 755 rows and 3 columns.
}
\usage{
canopy_cover_table
}
\description{
A dataframe with estimates of canopy cover proportion with bapa and tpa
observations from FIA data in the Sierra Nevada region
}
\keyword{datasets}
