% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster Landis summary data using kmeans clustering algorithm. See
\code{stats::\link[stats]{kmeans}} for details.}
\usage{
cluster_data(data, vars, n_clusters)
}
\arguments{
\item{data}{dataframe output of \code{\link{process_landis}}}

\item{vars}{character vector of variables to be used for grouping pixels
into clusters}

\item{n_clusters}{integer number of clusters}
}
\value{
list containing an object of class \code{kmeans} and the scale
parameters of the data passed to the kmeans function.
}
\description{
Cluster Landis summary data using kmeans clustering algorithm. See
\code{stats::\link[stats]{kmeans}} for details.
}
\examples{
cluster_data(
  data = test_landis,
  vars = attr(test_landis, "spp_vars"),
  n = 30
)
}
