% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_cover.R
\name{estimate_canopy_cover}
\alias{estimate_canopy_cover}
\title{Estimate canopy cover}
\usage{
estimate_canopy_cover(trees)
}
\arguments{
\item{trees}{dataframe of tree records for a set of pixels}
}
\value{
dataframe with estimate of canopy cover proportion by pixel
}
\description{
Estimate canopy cover
}
\examples{
cc <- estimate_canopy_cover(test_backcasted$trees)
}
