% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{get_matching_map_codes}
\alias{get_matching_map_codes}
\title{Get matching map_codes from lookup data}
\usage{
get_matching_map_codes(new_data, lookup, clusters)
}
\arguments{
\item{new_data}{dataframe object that needs nearest neighbor matches, output
of \code{\link{process_landis}}.}

\item{lookup}{dataframe output of \code{\link{process_landis}} from which
matches will be drawn.}

\item{clusters}{list output from \code{\link{cluster_data}}.}
}
\value{
dataframe of map_codes from \code{new_data} and matching map_codes
from \code{lookup}.
}
\description{
Get matching map_codes from lookup data
}
\examples{
# add cluster IDs to test_landis
test_landis$cluster <- test_clusters$kmeans$cluster
get_matching_map_codes(
  new_data = process_landis(data.table::fread(test_landis_prediction_file)),
  lookup = test_landis,
  clusters = test_clusters
)
}
