% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_code_crosswalk}
\alias{map_code_crosswalk}
\title{A dataframe that relates \code{map_code} values to 90 meter pixel
coordinates in EPSG:2163 projection.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1291995 rows and 4 columns.
}
\usage{
map_code_crosswalk
}
\description{
A dataframe that relates \code{map_code} values to 90 meter pixel
coordinates in EPSG:2163 projection.
}
\keyword{datasets}
