% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{process_landis}
\alias{process_landis}
\title{Process Landis summary data into wide dataframe for knn matching}
\usage{
process_landis(data)
}
\arguments{
\item{data}{dataframe containing observations of biomass (grams per sq. m) by
species (\code{landis_species}), age class (\code{age_class}), and pixel
(\code{map_code})}
}
\value{
wide dataframe with one row per pixel, a column with total biomass,
one column per species, and one column per age class
}
\description{
Process Landis summary data into wide dataframe for knn matching
}
\examples{
process_landis(data.table::fread(test_landis_training_file))
}
