% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_cover.R
\name{pull_canopy_cover}
\alias{pull_canopy_cover}
\title{Pull canopy cover estimate from FIA data}
\usage{
pull_canopy_cover(bapa, tpa, k = 5)
}
\arguments{
\item{bapa}{numeric vector of basal area (sq. ft per acre) in stems 4" and
larger.}

\item{tpa}{numeric vector of trees per acre in stems 4" and larger.}

\item{k}{integer number of matching records from which average canopy cover
estimate will be drawn}
}
\value{
numeric vector of canopy cover estimates expressed as a proportion
(0 - 1).
}
\description{
Pull canopy cover estimate from FIA data
}
\examples{
pull_canopy_cover(bapa = c(100, 200), tpa = c(100, 200), k = 5)
}
