% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{pull_matches}
\alias{pull_matches}
\title{Get nearest neighbor matches for new data from pool of pixels}
\usage{
pull_matches(new_data, lookup, vars)
}
\arguments{
\item{new_data}{dataframe object that needs nearest neighbor matches, output
of \code{\link{process_landis}}}

\item{lookup}{dataframe output of \code{\link{process_landis}} from which
matches will be drawn}

\item{vars}{character vector of variables to use for k nearest neighbor
matching}
}
\value{
dataframe of map_codes from \code{new_data} and matching map_codes
from \code{lookup}
}
\description{
Get nearest neighbor matches for new data from pool of pixels
}
\examples{
pull_matches(
  new_data = process_landis(data.table::fread(test_landis_prediction_file)),
  lookup = test_landis,
  vars = c(
    "aboveground_biomass_g_per_m2",
    attr(test_landis, "spp_vars"),
    attr(test_landis, "age_vars")
  )
)
}
