% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcaster.R
\name{summarize_matching_performance}
\alias{summarize_matching_performance}
\title{Summarize landis summary matching performance}
\usage{
summarize_matching_performance(match_frame, new_data, lookup)
}
\arguments{
\item{match_frame}{dataframe output from
\code{\link{get_matching_map_codes}} with \code{map_code} values for the
target pixels and for the matched records from the lookup dataset.}

\item{new_data}{dataframe object that needs nearest neighbor matches, output
of \code{\link{process_landis}}.}

\item{lookup}{dataframe output of \code{\link{process_landis}} from which
matches are drawn.}
}
\value{
list a containing a dataframe with matching accuracy statistics
(\code{comp_stats}), and a dataframe with original and matched values of
biomass by species and age class for \code{new_data} (\code{comp_frame})
}
\description{
Summarize landis summary matching performance
}
\examples{
summarize_matching_performance(
  match_frame = test_match_frame,
  new_data = process_landis(data.table::fread(test_landis_prediction_file)),
  lookup = test_landis
)
}
