% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_clusters}
\alias{test_clusters}
\title{The output of \code{\link{cluster_data}} when applied to
\code{\link{test_landis}} to generate 20 clusters based on biomass by
species.}
\format{
An object of class \code{list} of length 3.
}
\usage{
test_clusters
}
\description{
The output of \code{\link{cluster_data}} when applied to
\code{\link{test_landis}} to generate 20 clusters based on biomass by
species.
}
\keyword{datasets}
