% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_landis}
\alias{test_landis}
\title{The dataframe created by \code{\link{process_landis}} after reading
\code{\link{test_landis_training_file}}.}
\format{
An object of class \code{data.frame} with 2459 rows and 74 columns.
}
\usage{
test_landis
}
\description{
The dataframe created by \code{\link{process_landis}} after reading
\code{\link{test_landis_training_file}}.
}
\keyword{datasets}
