% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_landis_prediction_file}
\alias{test_landis_prediction_file}
\title{A compressed csv file containing observations of biomass by age class and
species to be used as a target dataset for testing backcasting.}
\format{
An object of class \code{character} of length 1.
}
\usage{
test_landis_prediction_file
}
\description{
A compressed csv file containing observations of biomass by age class and
species to be used as a target dataset for testing backcasting.
}
\keyword{datasets}
