% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_landis_treelist_dir}
\alias{test_landis_treelist_dir}
\title{The file path to a directory containing the compressed treelist csv that
corresponds to \code{\link{test_landis_training_file}}.}
\format{
An object of class \code{character} of length 1.
}
\usage{
test_landis_treelist_dir
}
\description{
The file path to a directory containing the compressed treelist csv that
corresponds to \code{\link{test_landis_training_file}}.
}
\keyword{datasets}
