% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_match_frame}
\alias{test_match_frame}
\title{The output of \code{\link{get_matching_map_codes}} when applied to
\code{\link{test_landis_prediction_file}} to identify nearest neighbor
matches from \code{\link{test_landis}}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2500 rows and 2 columns.
}
\usage{
test_match_frame
}
\description{
The output of \code{\link{get_matching_map_codes}} when applied to
\code{\link{test_landis_prediction_file}} to identify nearest neighbor
matches from \code{\link{test_landis}}.
}
\keyword{datasets}
